adding permessions

+++ tourism_in_Saudi/citys/views.py
@@ -4,8 +4,8 @@ from rest_framework.request import Request
 
 from rest_framework import status
 
-from .models import City, Place
-from .serializers import CitySerializer, PlaceSerializer
+from .models import City, Place, Comment
+from .serializers import CitySerializer, PlaceSerializer, CommentSerializer
 
 from rest_framework.decorators import authentication_classes, permission_classes
 from rest_framework_simplejwt.authentication import JWTAuthentication
@@ -24,6 +24,10 @@ def add_city(request: Request):
     if not user.is_authenticated:
         return Response({"msg" : "Please Log In"})
 
+    if not user.has_perm('citys.add_city'):
+        return Response({"msg" : "You don't have permission ! contact your admin"}, status=status.HTTP_401_UNAUTHORIZED)
+
+
     city_serializer = CitySerializer(data=request.data)
 
     if city_serializer.is_valid():
@@ -35,6 +39,8 @@ def add_city(request: Request):
 
 
 @api_view(["GET"])
+@authentication_classes([JWTAuthentication])
+@permission_classes([IsAuthenticated])
 def list_city(request: Request):
     citys = City.objects.all()
     citys_data = CitySerializer(instance=citys, many=True).data
@@ -43,7 +49,18 @@ def list_city(request: Request):
 
 
 @api_view(["PUT"])
+@authentication_classes([JWTAuthentication])
+@permission_classes([IsAuthenticated])
 def update_city(request: Request, city_id):
+
+    user = request.user
+
+    if not user.is_authenticated:
+        return Response({"msg" : "Please Log In"})
+
+    if not user.has_perm('citys.update_city'):
+        return Response({"msg" : "You don't have permission ! contact your admin"}, status=status.HTTP_401_UNAUTHORIZED)
+
     try:
         city = City.objects.get(id=city_id)
     except Exception as e:
@@ -60,7 +77,18 @@ def update_city(request: Request, city_id):
 
 
 @api_view(["DELETE"])
+@authentication_classes([JWTAuthentication])
+@permission_classes([IsAuthenticated])
 def delete_city(request: Request, city_id):
+    
+    user = request.user
+
+    if not user.is_authenticated:
+        return Response({"msg" : "Please Log In"})
+
+    if not user.has_perm('citys.delete_city'):
+        return Response({"msg" : "You don't have permission ! contact your admin"}, status=status.HTTP_401_UNAUTHORIZED)
+
     try:
         city = City.objects.get(id=city_id)
         city.delete()
@@ -73,6 +101,8 @@ def delete_city(request: Request, city_id):
 ### Place Views
 
 @api_view(["GET"])
+@authentication_classes([JWTAuthentication])
+@permission_classes([IsAuthenticated])
 def list_place(request: Request):
     places = Place.objects.all()
     places_data = PlaceSerializer(instance=places, many=True).data
@@ -90,6 +120,10 @@ def add_place(request: Request):
     if not user.is_authenticated:
         return Response({"msg" : "Please Log In"})
 
+    if not user.has_perm('citys.add_place'):
+        return Response({"msg" : "You don't have permission ! contact your admin"}, status=status.HTTP_401_UNAUTHORIZED)
+
+
     place_serializer = PlaceSerializer(data=request.data)
 
     if place_serializer.is_valid():
@@ -101,7 +135,18 @@ def add_place(request: Request):
 
 
 @api_view(["PUT"])
+@authentication_classes([JWTAuthentication])
+@permission_classes([IsAuthenticated])
 def update_place(request: Request, place_id):
+
+    user = request.user
+
+    if not user.is_authenticated:
+        return Response({"msg" : "Please Log In"})
+
+    if not user.has_perm('citys.update_place'):
+        return Response({"msg" : "You don't have permission ! contact your admin"}, status=status.HTTP_401_UNAUTHORIZED)
+
     try:
         place = Place.objects.get(id=place_id)
     except Exception as e:
@@ -118,7 +163,15 @@ def update_place(request: Request, place_id):
 
 
 @api_view(["DELETE"])
+@authentication_classes([JWTAuthentication])
+@permission_classes([IsAuthenticated])
 def delete_place(request: Request, place_id):
+
+    user = request.user
+
+    if not user.has_perm('citys.delete_place'):
+        return Response({"msg" : "You don't have permission ! contact your admin"}, status=status.HTTP_401_UNAUTHORIZED)
+
     try:
         place = Place.objects.get(id=place_id)
         place.delete()
@@ -127,3 +180,59 @@ def delete_place(request: Request, place_id):
 
     return Response({"msg": f"{place.name} has been deleted!!"})
 
+
+@api_view(['POST'])
+@authentication_classes([JWTAuthentication])
+@permission_classes([IsAuthenticated])
+def add_comment(request : Request):
+
+    user = request.user
+
+    if not user.is_authenticated:
+        return Response({"msg" : "Please Log In"})
+
+    if not user.has_perm('citys.add_comment'):
+        return Response({"msg" : "You don't have permission ! contact your admin"}, status=status.HTTP_401_UNAUTHORIZED)
+
+
+    comment_serializer = CommentSerializer(data=request.data)
+
+    if comment_serializer.is_valid():
+        comment_serializer.save()
+    else:
+        return Response({"msg" : "couldn't create a comment", "errors" : comment_serializer.errors}, status=status.HTTP_403_FORBIDDEN)
+
+    
+    return Response({"msg" : "Comment Added Successfully!"}, status=status.HTTP_201_CREATED)
+
+
+@api_view(['GET'])
+@authentication_classes([JWTAuthentication])
+@permission_classes([IsAuthenticated])
+def list_comments(request : Request):
+    comments = Comment.objects.all()
+    comments_data = CommentSerializer(instance=comments, many=True).data
+
+    return Response({"msg" : "list of all comments", "comments" : comments_data})
+
+
+@api_view(["DELETE"])
+@authentication_classes([JWTAuthentication])
+@permission_classes([IsAuthenticated])
+def delete_comment(request: Request, comment_id):
+
+    user = request.user
+
+    if not user.is_authenticated:
+        return Response({"msg" : "Please Log In"})
+
+    if not user.has_perm('citys.delete_comment'):
+        return Response({"msg" : "You don't have permission ! contact your admin"}, status=status.HTTP_401_UNAUTHORIZED)
+
+    try:
+        comment = Comment.objects.get(id = comment_id)
+        comment.delete()
+    except Exception as e:
+        return Response({"msg": "The comment is not found!"})
+
+    return Response({"msg": "the comment was deleted"})
